--- a/include/vrv/add.h
+++ b/include/vrv/add.h
@@ -29,7 +29,6 @@ public:
     virtual Object *Clone() const { return new Add(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Add"; }
-    virtual ClassId GetClassId() const { return ADD; }
     ///@}
 
 private: