--- a/include/vrv/anchoredtext.h
+++ b/include/vrv/anchoredtext.h
@@ -34,7 +34,6 @@ public:
     virtual Object *Clone() const { return new AnchoredText(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "AnchoredText"; }
-    virtual ClassId GetClassId() const { return ANCHOREDTEXT; }
     ///@}
 
     /**