--- a/include/vrv/arpeg.h
+++ b/include/vrv/arpeg.h
@@ -40,7 +40,6 @@ public:
     virtual Object *Clone() const { return new Arpeg(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Arpeg"; }
-    virtual ClassId GetClassId() const { return ARPEG; }
     ///@}
 
     /**