--- a/include/vrv/barline.h
+++ b/include/vrv/barline.h
@@ -32,11 +32,11 @@ public:
      */
     ///@{
     BarLine();
+    BarLine(ClassId classId);
     virtual ~BarLine();
     virtual Object *Clone() const { return new BarLine(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "BarLine"; }
-    virtual ClassId GetClassId() const { return BARLINE; }
     ///@}
 
     /** Override the method since alignment is required */
@@ -87,18 +87,15 @@ public:
     virtual ~BarLineAttr();
     virtual Object *Clone() const { return new BarLineAttr(*this); }
     virtual std::string GetClassName() const { return "BarLineAttr"; }
-    virtual ClassId GetClassId() const
-    {
-        if (m_noAttr) return BARLINE;
-        return m_isLeft ? BARLINE_ATTR_LEFT : BARLINE_ATTR_RIGHT;
-    }
     ///@}
 
-    void SetLeft() { m_isLeft = true; }
-    void SetNoAttr() { m_noAttr = true; }
+    void SetLeft();
+    void SetNoAttr();
 
 private:
-    //
+    /** The class id here depends on the value of the class members */
+    void UpdateClassId();
+
 public:
     //
 private: