--- a/include/vrv/bracketspan.h
+++ b/include/vrv/bracketspan.h
@@ -38,7 +38,6 @@ public:
     virtual Object *Clone() const { return new BracketSpan(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "BracketSpan"; }
-    virtual ClassId GetClassId() const { return BRACKETSPAN; }
     ///@}
 
     /**