--- a/include/vrv/breath.h
+++ b/include/vrv/breath.h
@@ -33,7 +33,6 @@ public:
     virtual Object *Clone() const { return new Breath(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Breath"; }
-    virtual ClassId GetClassId() const { return BREATH; }
     ///@}
 
     /**