--- a/include/vrv/btrem.h
+++ b/include/vrv/btrem.h
@@ -32,7 +32,6 @@ public:
     virtual Object *Clone() const { return new BTrem(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "BTrem"; }
-    virtual ClassId GetClassId() const { return BTREM; }
     ///@}
 
     /**