--- a/include/vrv/clef.h
+++ b/include/vrv/clef.h
@@ -41,7 +41,6 @@ public:
     virtual Object *Clone() const { return new Clef(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Clef"; }
-    virtual ClassId GetClassId() const { return CLEF; }
     ///@}
 
     /** Override the method since alignment is required */