--- a/include/vrv/corr.h
+++ b/include/vrv/corr.h
@@ -29,7 +29,6 @@ public:
     virtual Object *Clone() const { return new Corr(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Corr"; }
-    virtual ClassId GetClassId() const { return CORR; }
     ///@}
 
 private: