--- a/include/vrv/custos.h
+++ b/include/vrv/custos.h
@@ -33,7 +33,6 @@ public:
     virtual Object *Clone() const { return new Custos(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Custos"; }
-    virtual ClassId GetClassId() const { return CUSTOS; }
     ///@}
 
     /**