--- a/include/vrv/damage.h
+++ b/include/vrv/damage.h
@@ -29,7 +29,6 @@ public:
     virtual Object *Clone() const { return new Damage(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Damage"; }
-    virtual ClassId GetClassId() const { return DAMAGE; }
     ///@}
 
 private: