--- a/include/vrv/del.h
+++ b/include/vrv/del.h
@@ -29,7 +29,6 @@ public:
     virtual Object *Clone() const { return new Del(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Del"; }
-    virtual ClassId GetClassId() const { return DEL; }
     ///@}
 
 private: