--- a/include/vrv/dir.h
+++ b/include/vrv/dir.h
@@ -42,7 +42,6 @@ public:
     virtual Object *Clone() const { return new Dir(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Dir"; }
-    virtual ClassId GetClassId() const { return DIR; }
     ///@}
 
     /**