--- a/include/vrv/doc.h
+++ b/include/vrv/doc.h
@@ -46,7 +46,6 @@ public:
     ///@{
     Doc();
     virtual ~Doc();
-    virtual ClassId GetClassId() const { return DOC; }
     ///@}
 
     /**