--- a/include/vrv/dot.h
+++ b/include/vrv/dot.h
@@ -31,7 +31,6 @@ public:
     virtual Object *Clone() const { return new Dot(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Dot"; }
-    virtual ClassId GetClassId() const { return DOT; }
     ///@}
 
     /**