--- a/include/vrv/editorial.h
+++ b/include/vrv/editorial.h
@@ -43,10 +43,10 @@ public:
      */
     ///@{
     EditorialElement();
-    EditorialElement(const std::string &classid);
+    EditorialElement(ClassId classId);
+    EditorialElement(ClassId classId, const std::string &classIdStr);
     virtual ~EditorialElement();
     virtual void Reset();
-    virtual ClassId GetClassId() const { return EDITORIAL_ELEMENT; }
     ///@}
 
     /**