--- a/include/vrv/expansion.h
+++ b/include/vrv/expansion.h
@@ -35,7 +35,6 @@ public:
     virtual Object *Clone() const { return new Expansion(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Expansion"; }
-    virtual ClassId GetClassId() const { return EXPANSION; }
     ///@}
 
     /**