--- a/include/vrv/facsimile.h
+++ b/include/vrv/facsimile.h
@@ -39,7 +39,6 @@ public:
     virtual ~Facsimile();
     virtual Object *Clone() const { return new Facsimile(*this); }
     virtual void Reset();
-    virtual ClassId GetClassId() const { return FACSIMILE; }
     virtual std::string GetClassName() const { return "facsimile"; }
     ///@}
     virtual bool IsSupportedChild(Object *object);