--- a/include/vrv/fing.h
+++ b/include/vrv/fing.h
@@ -34,7 +34,6 @@ public:
     virtual Object *Clone() const { return new Fing(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Fing"; }
-    virtual ClassId GetClassId() const { return FING; }
     ///@}
 
     /**