--- a/include/vrv/ftrem.h
+++ b/include/vrv/ftrem.h
@@ -34,7 +34,6 @@ public:
     virtual Object *Clone() const { return new FTrem(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "FTrem"; }
-    virtual ClassId GetClassId() const { return FTREM; }
     ///@}
 
     /**