--- a/include/vrv/gracegrp.h
+++ b/include/vrv/gracegrp.h
@@ -29,7 +29,6 @@ public:
     virtual Object *Clone() const { return new GraceGrp(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "GraceGrp"; };
-    virtual ClassId GetClassId() const { return GRACEGRP; };
     ///@}
 
     /**