--- a/include/vrv/halfmrpt.h
+++ b/include/vrv/halfmrpt.h
@@ -35,7 +35,6 @@ public:
     virtual Object *Clone() const { return new HalfmRpt(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "HalfmRpt"; }
-    virtual ClassId GetClassId() const { return HALFMRPT; }
     ///@}
 
     /** Override the method since alignment is required */