--- a/include/vrv/lb.h
+++ b/include/vrv/lb.h
@@ -32,7 +32,6 @@ public:
     virtual Object *Clone() const { return new Lb(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Lb"; }
-    virtual ClassId GetClassId() const { return LB; }
     ///@}
 
     /**