--- a/include/vrv/lem.h
+++ b/include/vrv/lem.h
@@ -29,7 +29,6 @@ public:
     virtual Object *Clone() const { return new Lem(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Lem"; }
-    virtual ClassId GetClassId() const { return LEM; }
     ///@}
 
 private: