--- a/include/vrv/mdiv.h
+++ b/include/vrv/mdiv.h
@@ -34,7 +34,6 @@ public:
     virtual Object *Clone() const { return new Mdiv(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "Mdiv"; }
-    virtual ClassId GetClassId() const { return MDIV; }
     ///@}
 
     /**