--- a/include/vrv/mnum.h
+++ b/include/vrv/mnum.h
@@ -41,7 +41,6 @@ public:
     virtual Object *Clone() const { return new MNum(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "MNum"; }
-    virtual ClassId GetClassId() const { return MNUM; }
     ///@}
 
     /**