--- a/include/vrv/mrest.h
+++ b/include/vrv/mrest.h
@@ -41,7 +41,6 @@ public:
     virtual Object *Clone() const { return new MRest(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "MRest"; }
-    virtual ClassId GetClassId() const { return MREST; }
     ///@}
 
     /**