--- a/include/vrv/mrpt.h
+++ b/include/vrv/mrpt.h
@@ -35,7 +35,6 @@ public:
     virtual Object *Clone() const { return new MRpt(*this); }
     virtual void Reset();
     virtual std::string GetClassName() const { return "MRpt"; }
-    virtual ClassId GetClassId() const { return MRPT; }
     ///@}
 
     //----------//