--- a/include/vrv/page.h
+++ b/include/vrv/page.h
@@ -40,7 +40,6 @@ public:
     virtual ~Page();
     virtual void Reset();
     virtual std::string GetClassName() const { return "Page"; }
-    virtual ClassId GetClassId() const { return PAGE; }
     ///@}
 
     /**