This PR should improve total runtime by ~10%. The idea is not to retrieve class IDs via virtual functions, but to store them at construction in the Object base class.

Some examples that I have tested:

| Piece | Size | Runtime before | Runtime after |
| ----- | ---- | ---------------- | ------------- |
| Beethoven - Presto | 1.1 MB / 10 pages | 530ms | 475ms |
| Hummel - Trumpet Concerto | 5.7 MB / 128 pages | 4.52s | 3.94s |
| Holst - The Planets | 50.1 MB / 403 pages | 2min 8s | 1min 56s |