@lpugin 2021-08-20
Looks good. My only concern is the introduction to Object::SetClassId because we should never have to change a ClassId. As far as I can see, this is only for BarLineAttr. Am I right? I think it would be better to have two distinct BarLineAttrLeft/Right classes so we can forget it. Can you look at this?