@DavidBauer1984 2021-08-20
Actually I tried it first this way, but then I got worried, because `BarLineAttr` also impersonates itself as 'BarLine' when 'm_noAttr' is set. But I can have a look again and agree that it would be much nicer without Object::SetClassId.