@DavidBauer1984 2021-08-20
@lpugin What do you think about having a bool option in Verovio --show-runtime that displays total runtime if desired. I would find it helpful as developer to keep an eye on runtime without having to launch instruments in Xcode (or other tools).