@lpugin 2021-08-20
This makes me see that this is actually not good https://github.com/rism-digital/verovio/commit/2b909f61e50ca465ceba13b637d02012cd722106 and should probably not have made its way through.
ClassIds were introduced for speed optimization. The idea is that we do not do `dynamic_cast` of object type identification. Since now we even changed to `static_cast` (through `vrv_cast`) we should absolutely forbid ClassId to be changed.

Maybe we don't need BarlLineAttr at all and only rely on some flags in `BarLine`. I cannot remember why having `BarLineAttr` was the simplest option when it was introduced.