@DavidBauer1984 2021-08-23
@rettinghaus You are right, I have reverted the change. This was the remnant of my fix for #2344 which has already been fixed by Laurent from another direction. So it has no big effect on runtime anymore, even though it still means a lot of unnecessary iteration, but we can optimize this another time.

@lpugin I have removed the BarLineAttr, but it was not straightforward. I guess the main reasons why it was introduced was to have another class ID to distinguish in:

- Doc::GetLeftMargin(ClassID id) / Doc::GetRightMargin(ClassID id)
- The class ID include list for the AdjustXPos functor.

To preserve the distinction I have added an additional overload in the first case and an additional param in the second case.